+++ src/finder/mod.rs
@@ -1,2 +1,3 @@
 pub mod example;
+pub mod mapper;
 pub mod unsafe_finder;
