+++ src/lib.rs
@@ -12,6 +12,7 @@ extern crate rustc_errors;
 extern crate rustc_feature;
 extern crate rustc_hash;
 extern crate rustc_hir;
+extern crate rustc_hir_pretty;
 extern crate rustc_index;
 extern crate rustc_interface;
 extern crate rustc_literal_escaper;
@@ -28,6 +29,7 @@ extern crate thin_vec;
 pub mod preprocessor;
 #[macro_use]
 pub mod ast_util;
+pub mod ast_hir;
 pub mod bin_file_adder;
 pub mod bit_set;
 pub mod check_unsafety;
